feat: let a chapter row open reading mode directly
Clicking a chapter meant landing on its page and then pressing "Read this
chapter" — an interstitial for anyone browsing a book in order to read it.

Each row now carries two destinations. The title still goes to the chapter's own
page, because that is the crawlable, rankable URL and where a search engine sends
people; the icon beside it opens reading mode scoped to that chapter. Pointing
the title at reading mode instead would have been fewer clicks but would spend
the book and volume pages' strongest internal links on URLs that are noindex and
client-rendered, leaving the 7,700 chapter pages to rank on breadcrumbs alone.

The icon is faint rather than hidden until hover: an affordance that only appears
on hover is one most readers never find, and on touch it would never appear at
all.

readingModeUrl moves from the controller onto BookCatalog so a Chapter can hand
out its own reading URL, rather than the page having to build one per row.
